Refactor: 변경된 파일구조 적용
